Реализованы права в EmployeeService #12 Тесты Employee дополнены правами #13 Реализованы тесты для сущности  staff #10

+++ Data/Catdog50RUS.EmployeesAccountingSystem.Data/Services/EmployeeService/EmployeeService.cs
@@ -1,4 +1,5 @@
 ﻿using Catdog50RUS.EmployeesAccountingSystem.Data.Repository;
+using Catdog50RUS.EmployeesAccountingSystem.Models;
 using Catdog50RUS.EmployeesAccountingSystem.Models.Employees;
 using System;
 using System.Collections.Generic;
@@ -7,22 +8,26 @@ using System.Threading.Tasks;
 
 namespace Catdog50RUS.EmployeesAccountingSystem.Data.Services.EmployeeService
 {
+
     public class EmployeeService : IEmployeeService
     {
         /// <summary>
         /// Внедрение зависимости через интерфейс
         /// </summary>
         private readonly IEmployeeRepository _employeeRepository;
+        private readonly Autorize _autorize;
 
         public bool IsFirstRun { get; }
 
         /// <summary>
         /// Конструктор
         /// </summary>
-        public EmployeeService(IEmployeeRepository repository)
+        public EmployeeService(IEmployeeRepository repository, Autorize autorize)
         {
             _employeeRepository = repository;
-            //IsFirstRun = _employeeRepository.IsFirstRun;
+
+            if (autorize.UserRole == Role.Director)
+                _autorize = autorize;
         }
 
         #region Interface
@@ -34,6 +39,9 @@ namespace Catdog50RUS.EmployeesAccountingSystem.Data.Services.EmployeeService
         /// <returns></returns>
         public async Task<bool> InsertEmployeeAsync(BaseEmployee employee)
         {
+            //Проверяем права доступа
+            if (_autorize == null)
+                return false;
             //Проверяем входные параметры на null
             if (employee == null)
                 return false;
@@ -60,6 +68,9 @@ namespace Catdog50RUS.EmployeesAccountingSystem.Data.Services.EmployeeService
         /// <returns></returns>
         public async Task<IEnumerable<BaseEmployee>> GetAllEmployeeAsync()
         {
+            //Проверяем права доступа
+            if (_autorize == null)
+                return null;
             return await _employeeRepository.GetEmployeesListAsync();
         }
 
@@ -70,6 +81,9 @@ namespace Catdog50RUS.EmployeesAccountingSystem.Data.Services.EmployeeService
         /// <returns></returns>
         public async Task<BaseEmployee> GetEmployeeByNameAsync(string name)
         {
+            //Проверяем права доступа
+            if (_autorize == null)
+                return null;
             //Проверяем входной параметр на пустоту и null
             if (string.IsNullOrWhiteSpace(name))
                 return null;
@@ -84,6 +98,9 @@ namespace Catdog50RUS.EmployeesAccountingSystem.Data.Services.EmployeeService
         /// <returns></returns>
         public async Task<bool> DeleteEmployeeAsync(Guid id)
         {
+            //Проверяем права доступа
+            if (_autorize == null)
+                return false;
             //Пробуем удалить сотрудника из хранилища
             var result = await _employeeRepository.DeleteEmployeeAsync(id);
             //Если результат не null возвращаем true, иначе false
@@ -97,6 +114,9 @@ namespace Catdog50RUS.EmployeesAccountingSystem.Data.Services.EmployeeService
         /// </summary>
         public async Task<bool> DeleteEmployeeByNameAsync(string name)
         {
+            //Проверяем права доступа
+            if (_autorize == null)
+                return false;
             //Проверяем имя на пустое значение или null
             if (string.IsNullOrWhiteSpace(name))
                 return false;
